requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
channels~=2.4.0
channels-redis==2.4.2
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.6
daphne==2.5.0
Django==3.1.7
django-js-asset==1.2.2
django-mptt==0.11.0
django-redis==4.12.1
et-xmlfile==1.0.1
hiredis==1.1.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
jdcal==1.4.1
Jinja2==2.11.3
lxml==4.6.2
MarkupSafe==1.1.1
msgpack==0.6.2
openpyxl==3.0.6
Pillow==8.1.0
prettytable==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyecharts==1.9.0
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
urllib3==1.26.3
wcwidth==0.2.5
zope.interface==5.2.0

numpy~=1.19.2
scipy~=1.2.1
torchvision~=0.9.0
matplotlib~=3.3.4
scikit-image~=0.17.2

h5py~=2.10.0
tqdm~=4.59.0
